feat(engine): add zip functionality and compressOutput option to action schema
