+++ engine/runtime/action-sink/src/file/cesium3dtiles.rs
@@ -1,3 +1,4 @@
+mod pipeline;
 pub(crate) mod sink;
 mod slice;
 mod tiling;
