+++ engine/schema/actions_es.json
@@ -615,6 +615,16 @@
               "null"
             ]
           },
+          "compressOutput": {
+            "anyOf": [
+              {
+                "$ref": "#/definitions/Expr"
+              },
+              {
+                "type": "null"
+              }
+            ]
+          },
           "maxZoom": {
             "type": "integer",
             "format": "uint8",
